l2o: address review nits and add QF_NIA L2O regressions
- terms: make get_composite's shared scratch buffer thread-local and document
  that its returned descriptor is only valid until the next get_composite call
  (the buffer was promoted from preprocessor.c to a public terms.h API).
- trail: drop the dead trail_model_cache_clear declaration (never defined).
- mcsat/solver: remove the duplicate <math.h> include.
- l2o: fix signed/unsigned loop indices (use uint32_t / size_t against
  c->arity, n, state->n_var and fs->size) in l2o.c, evaluator.c, hill_climbing.c.
- l2o_cost_fx.c: add the missing GPL license header.
- tests: add two QF_NIA regressions (sum-of-squares) under mcsat/nia that
  actually exercise the L2O path (l2o::runs >= 1, ~0.05s each, both sat).

commit supported by Windsurf/Opus4.7

+++ src/mcsat/l2o/hill_climbing.c
@@ -198,7 +198,7 @@ bool optimize_fs(l2o_t *l2o, l2o_cost_fx_t *fx, l2o_search_state_t *state, uint3
 
   lp_value_t lp_val;
   lp_value_construct_zero(&lp_val);
-  for (int i = 0; i < fs->size; ++i) {
+  for (size_t i = 0; i < fs->size; ++i) {
     const lp_interval_t *interval = &fs->intervals[i];
     lp_interval_pick_value(interval, &lp_val);
     *val = lp_value_to_double(&lp_val);
